Geometric search update
Former-commit-id: 472bfd99e51dd5fb301832d2b7956b5154381ca0

+++ ConsoleApplication/GeometricSearchModule/src/org/janelia/it/workstation/gui/geometric_search/viewer/actor/DenseVolumeActor.java
@@ -12,13 +12,13 @@ import java.util.List;
 public class DenseVolumeActor extends Actor {
 
     List<Vector4> voxels;
-    int xSize;
-    int ySize;
-    int zSize;
+    float xSize;
+    float ySize;
+    float zSize;
     float voxelUnitSize;
     ArrayCubeGLActor arrayCubeGLActor;
 
-    public DenseVolumeActor(String name, List<Vector4> voxels, int xSize, int ySize, int zSize, float voxelUnitSize) {
+    public DenseVolumeActor(String name, List<Vector4> voxels, float xSize, float ySize, float zSize, float voxelUnitSize) {
         this.name=name;
         this.voxels=voxels;
         this.xSize=xSize;
